+++ src/GitHub_Updater/API.php
@@ -63,12 +63,20 @@ class API {
 	protected static $options;
 
 	/**
+	 * Holds extra headers.
+	 *
+	 * @var
+	 */
+	protected static $extra_headers;
+
+	/**
 	 * API constructor.
 	 *
 	 */
 	public function __construct() {
-		$this->base      = $base = Singleton::get_instance( 'Base' );
-		static::$options = $base::$options;
+		$this->base            = $base = Singleton::get_instance( 'Base' );
+		static::$options       = $base::$options;
+		static::$extra_headers = $this->base->add_headers( array() );
 	}
 
 	/**
@@ -650,7 +658,6 @@ class API {
 		$this->type->remote_version       = strtolower( $response['Version'] );
 		$this->type->requires_php_version = ! empty( $response['Requires PHP'] ) ? $response['Requires PHP'] : $this->type->requires_php_version;
 		$this->type->requires_wp_version  = ! empty( $response['Requires WP'] ) ? $response['Requires WP'] : $this->type->requires_wp_version;
-		$this->type->release_asset        = ( ! empty( $response['Release Asset'] ) && 'true' === $response['Release Asset'] );
 		$this->type->dot_org              = $response['dot_org'];
 	}