+++ src/GitHub_Updater/Base.php
@@ -48,7 +48,7 @@ class Base {
 	 *
 	 * @var array
 	 */
-	protected static $extra_headers = array();
+	public static $extra_headers = array();
 
 	/**
 	 * Holds the values to be used in the fields callbacks.
@@ -393,22 +393,18 @@ class Base {
 	 * @return array
 	 */
 	public function add_headers( $extra_headers ) {
-		$uri_type          = '';
 		$ghu_extra_headers = array(
 			'Requires WP'   => 'Requires WP',
 			'Requires PHP'  => 'Requires PHP',
 			'Release Asset' => 'Release Asset',
 		);
 
-		$current_filter = current_filter();
-		if ( 'extra_plugin_headers' === $current_filter ) {
-			$uri_type = ' Plugin URI';
-		} elseif ( 'extra_theme_headers' === $current_filter ) {
-			$uri_type = ' Theme URI';
-		}
+		$uri_types = array( 'plugin' => ' Plugin URI', 'theme' => ' Theme URI' );
 
 		foreach ( self::$git_servers as $server ) {
-			$ghu_extra_headers[ $server . $uri_type ] = $server . $uri_type;
+			foreach ( $uri_types as $uri_type ) {
+				$ghu_extra_headers[ $server . $uri_type ] = $server . $uri_type;
+			}
 			foreach ( self::$extra_repo_headers as $header ) {
 				$ghu_extra_headers[ $server . ' ' . $header ] = $server . ' ' . $header;
 			}
@@ -460,8 +456,6 @@ class Base {
 		$this->$type->open_issues          = 0;
 		$this->$type->requires_wp_version  = '4.6';
 		$this->$type->requires_php_version = '5.3';
-		$this->$type->release_asset        = false;
-		$this->$type->ci_job               = false;
 	}
 
 	/**
@@ -533,8 +527,7 @@ class Base {
 	private function is_cron_overdue( $cron, $timestamp ) {
 		$overdue = ( ( time() - $timestamp ) / HOUR_IN_SECONDS ) > 24;
 		if ( $overdue ) {
-			/* translators: 'git API check event' refers to the repository check against GitHub's (Bitbucket, GitLab) API */
-			$error_msg = esc_html__( 'There may be a problem with WP-Cron. The git server API check event is overdue.', 'github-updater' );
+			$error_msg = esc_html__( 'There may be a problem with WP-Cron. A GitHub Updater WP-Cron event is overdue.', 'github-updater' );
 			$error     = new \WP_Error( 'github_updater_cron_error', $error_msg );
 			Singleton::get_instance( 'Messages' )->create_error_message( $error );
 		}